refactor : 자동차 데이터(이름, 거리 등)를 가지는 data 패키지 생성
