--- src/main/java/racingcar/domain/Position.java
@@ -1,11 +1,0 @@
-package racingcar.domain;
-
-public class Position {
-    private int value;
-
-    public Position(int value) {
-        this.value = value;
-    }
-
-
-}
