+++ src/main/java/racingcar/domain/data/Name.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.data;
 
 import racingcar.utils.exception.InvalidNameLengthException;
